requirements.txt
pillow==10.3.0
plexapi==4.15.12
requests==2.31.0
python-dotenv==1.1.1
